feat(RenameTypes): remove extra result traversal
RenameTypes transform now relies on wrapping the resolveType methods within the outer schema instead of using transformResult. Required changing the resolveFromParentTypename to return a string instead of an object, as the string may need to be renamed before it corresponds to a type in the outer schema.

So that the transform can be used as a operation transform as well, the transformResult function is preserved, but a new flag is used to skip result transformation when the resolvers will be responsible for transforming the result, i.e. when used with transformSchema/wrapSchema.